requirements.txt
absl-py==2.3.1
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
argbind==0.3.9
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
bitsandbytes==0.46.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
descript-audio-codec==1.0.0
descript-audiotools @ git+https://github.com/descriptinc/audiotools@348ebf2034ce24e2a91a553e3171cb00c0c71678
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
executing==2.2.0
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
flash_attn==2.8.0.post2
flatten-dict==0.4.2
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
grpcio==1.73.1
hf-xet==1.1.5
huggingface-hub==0.33.2
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.5.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.0
parler_tts @ git+https://github.com/huggingface/parler-tts.git@d108732cd57788ec86bc857d99a6cabd66663d68
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.3
pystoi==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
randomname==0.2.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SpeechRecognition==3.14.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.7.1
torch-stoi==0.2.3
torchaudio==2.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
